backend/requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
channels==3.0.4
